Fix bf16-autocast crash in null-token scatter; loader type hygiene
The CFG null_token is a fp32 Parameter, but under bf16 autocast the projected
text tokens are bf16. An in-place scatter (text_tokens[empty,0] = null_token)
has no autocast cast hook, so it dtype-mismatches and crashes on CUDA. Match the
destination dtype. Every local test passed because MPS/CPU run autocast-off; add
a CPU bf16-autocast regression test that reproduces the GPU-only crash.

Also make the dataset classes subclass torch Dataset and tidy the transform list
so pyrefly stays at 0 errors.

+++ imagegen/text/conditioner.py
@@ -383,7 +383,9 @@ class TextConditioner(nn.Module):
             mask[empty] = False
             mask[empty, 0] = True
             text_tokens[empty] = 0
-            text_tokens[empty, 0] = self.null_token
+            # null_token is a fp32 Parameter; under bf16 autocast text_tokens is
+            # bf16 and an in-place scatter has no autocast cast hook, so match dtypes.
+            text_tokens[empty, 0] = self.null_token.to(text_tokens.dtype)
 
         return TextCondition(tokens=text_tokens, mask=mask)
